Extract scripts and reorganisation
